hw: vfio: quirks: Extract IGD code into its own file
We want to be able to compile out all or part of the VFIO PCI quirks, so
we need to put each and everyone of them into their own file.
We start with the IGD code.
