+++ hw/vfio/quirks/Makefile.objs
@@ -1,1 +1,1 @@
-obj-$(CONFIG_PCI) += pci.o
+obj-$(CONFIG_PCI) += pci.o pci-igd.o
